Moved account app api to api app
