+++ api/serializers.py
@@ -3,17 +3,22 @@ from rest_framework.response import Response
 
 from accounts.models import User
 from blog.models import Article
+from trips.models import Trip
 
-from blog.serializers import ArticleSerializer
 
+# accounts serializers
 
 class UserSerializer(serializers.ModelSerializer):
     likes_number = serializers.SerializerMethodField()
     articles = serializers.SerializerMethodField()
+    trips = serializers.SerializerMethodField()
 
     def get_articles(self, obj):
         return ArticleSerializer(Article.objects.filter(
             liked_by__in=[obj]), many=True).data
+
+    def get_trips(self, obj):
+        return Trip.objects.filter(driver_id=obj.id).count()
 
     def get_likes_number(self, obj):
         return Article.objects.filter(liked_by__in=[obj]).count()
@@ -28,7 +33,8 @@ class UserSerializer(serializers.ModelSerializer):
             'photo',
             'birthday',
             'likes_number',
-            'articles'
+            'articles',
+            'trips',
         )
 
 
@@ -194,3 +200,78 @@ class UserMainInfoSerializer(serializers.Serializer):
     class Meta:
         model = User
         fields = ('email', 'password')
+
+
+# blog serializers
+class ArticleSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Article
+        fields = ('id', 'title')
+
+
+class ArticleFullDataSerializer(serializers.ModelSerializer):
+    """
+    Class based on Article model and describes the
+    interface with full information about article:
+    title, body, data, author, liked_by.
+    """
+
+    class Meta:
+        model = Article
+        fields = ['title', 'body', 'author', 'liked_by']
+
+
+class ArticleCreateSerializer(serializers.ModelSerializer):
+    """
+    Class based on Article model and describes the
+    interface for creating new article.
+    User can write only article title and body,
+    author and data will add automaticaly.
+    """
+    class Meta:
+        model = Article
+        fields = ['title', 'body']
+
+    def validate(self, attrs):
+        # check the title name for uniqueness
+        if Article.objects.filter(title__exact=attrs.get('title')).exists():
+            raise serializers.ValidationError(
+                'Sorry, article with this title already exist. Please rename your article title'
+            )
+        return attrs
+
+
+class ArticleRemoveSerializer(serializers.ModelSerializer):
+    """
+    Class based on Article model and describes the interface for removing
+    article by title,
+    I suppose user know only this article parameter and don't see article id.
+    """
+    title = serializers.CharField()
+
+    class Meta:
+        model = Article
+        fields = ['title']
+
+    def validate(self, attrs):
+        # check if the title is exist
+        if not Article.objects.filter(title__exact=attrs.get('title')).exists():
+            raise serializers.ValidationError(
+                'Sorry, but article with title {} not exist'.format(
+                    attrs.get('title')))
+        return attrs
+
+
+class ArticleSearchSerializer(serializers.Serializer):
+    """
+    Class based on Serializer model and describes the interface for searching
+    article by some words or
+    letters in article title or body.
+    """
+    search_keyword = serializers.CharField()
+
+# trip serializers
+class TripSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Trip
+        fields = ('id', 'from_city')
